requirements.txt
django==3.2.9
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
psycopg2==2.9.2
gunicorn==20.1.0
django-heroku==0.3.1
django-cors-headers
